Move tests to sampling test
